Move plugin loader from TrackerInfo to ConfigStandalone
