+++ RecoTracker/MkFitCMS/standalone/buildtestMPlex.cc
@@ -5,7 +5,7 @@
 #include "RecoTracker/MkFitCMS/interface/MkStdSeqs.h"
 #include "RecoTracker/MkFitCMS/standalone/MkStandaloneSeqs.h"
 
-#include "tbb/parallel_for.h"
+#include "oneapi/tbb/parallel_for.h"
 
 #include <memory>
