+++ RecoTracker/MkFitCore/src/TrackerInfo.cc
@@ -98,68 +98,4 @@ namespace mkfit {
       return l2 == i1.m_sibl_barrel;
   }
 
-  //==============================================================================
-  // Plugin Loader
-  //==============================================================================
-
-#include <dlfcn.h>
-#include <sys/stat.h>
-#include <cstdlib>
-
-  namespace {
-    const char *search_path[] = {"", "../Geoms/", "Geoms/", "../", nullptr};
-  }
-
-  void TrackerInfo::execTrackerInfoCreatorPlugin(const std::string &base,
-                                                 TrackerInfo &ti,
-                                                 IterationsInfo &ii,
-                                                 bool verbose) {
-    std::string soname = base + ".so";
-
-    struct stat st;
-
-    int si = 0;
-    while (search_path[si]) {
-      std::string path;
-      const char *envpath = std::getenv("MKFIT_BASE");
-      if (envpath != nullptr) {
-        path += envpath;
-        path += "/";
-      }
-      path += search_path[si];
-      path += soname;
-      if (stat(path.c_str(), &st) == 0) {
-        printf("TrackerInfo::ExecTrackerInfoCreatorPlugin processing '%s'\n", path.c_str());
-
-        void *h = dlopen(path.c_str(), RTLD_LAZY);
-        if (!h) {
-          perror("dlopen failed");
-          exit(2);
-        }
-
-        long long *p2f = (long long *)dlsym(h, "TrackerInfoCrator_ptr");
-        if (!p2f) {
-          perror("dlsym failed");
-          exit(2);
-        }
-
-        TrackerInfoCreator_foo foo = (TrackerInfoCreator_foo)(*p2f);
-        foo(ti, ii, verbose);
-
-        // XXXXMT4Dan
-        // With dlclose I saw (rarely) valgrind errors saying
-        // Invalid read of size 4
-        //     at 0x4119F7: is_barrel (TrackerInfo.h:54)
-        // Commenting it out for now.
-        // dlclose(h);
-        return;
-      }
-
-      ++si;
-    }
-
-    fprintf(stderr, "TrackerInfo plugin '%s' not found in search path.\n", soname.c_str());
-    exit(2);
-  }
-
 }  // end namespace mkfit
